requirements.txt
Flask
python-dotenv
Werkzeug
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth 
pinecone-client
openai
tiktoken